+++ packages/solutions/monorepo-tools/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @modern-js/monorepo-tools
 
+<<<<<<< HEAD
 ## 1.20.1
 
 ### Patch Changes
@@ -34,6 +35,8 @@
   - @modern-js/upgrade@1.20.0
   - @modern-js/plugin@1.20.0
 
+=======
+>>>>>>> 6ab74256f (Release v1.19.0 (#1791))
 ## 1.19.0
 
 ### Patch Changes